Convert geometry to use 32-bit indices

+++ src/collider/GeomTree.h
@@ -19,7 +19,7 @@ struct BVHNode;
 
 class GeomTree {
 public:
-	GeomTree(const int numVerts, const int numTris, const std::vector<vector3f> &vertices, const Uint16 *indices, const unsigned int *triflags);
+	GeomTree(const int numVerts, const int numTris, const std::vector<vector3f> &vertices, const Uint32 *indices, const Uint32 *triflags);
 	GeomTree(Serializer::Reader &rd);
 	~GeomTree();
 
@@ -33,7 +33,7 @@ public:
 	//void TraceCoherentRays(int numRays, const vector3f &a_origin, const vector3f *a_dirs, isect_t *isects) const;
 	//void TraceCoherentRays(const BVHNode *startNode, int numRays, const vector3f &a_origin, const vector3f *a_dirs, isect_t *isects) const;
 	vector3f GetTriNormal(int triIdx) const;
-	unsigned int GetTriFlag(int triIdx) const { return m_triFlags[triIdx]; }
+	Uint32 GetTriFlag(int triIdx) const { return m_triFlags[triIdx]; }
 	double GetRadius() const { return m_radius; }
 	struct Edge {
 		int v1i, v2i;
@@ -67,8 +67,8 @@ public:
 	BVHTree* GetEdgeTree() const { return m_edgeTree.get(); }
 
 	const std::vector<vector3f>& GetVertices() const { return m_vertices; }
-	const Uint16 *GetIndices() const { return &m_indices[0]; }
-	const unsigned int *GetTriFlags() const { return &m_triFlags[0]; }
+	const Uint32 *GetIndices() const { return &m_indices[0]; }
+	const Uint32 *GetTriFlags() const { return &m_triFlags[0]; }
 	int GetNumVertices() const { return m_numVertices; }
 	int GetNumTris() const { return m_numTris; }
 
@@ -91,8 +91,8 @@ private:
 	std::vector<Edge> m_edges;
 
 	std::vector<vector3f> m_vertices;
-	std::vector<Uint16> m_indices;
-	std::vector<unsigned int> m_triFlags;
+	std::vector<Uint32> m_indices;
+	std::vector<Uint32> m_triFlags;
 };
 
 #endif /* _GEOMTREE_H */
